+++ src/main/java/com/msduoc/reservadehoteles/service/ReservaServiceImpl.java
@@ -1,29 +1,73 @@
 package com.msduoc.reservadehoteles.service;
 
+import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.msduoc.reservadehoteles.enums.EstadoReserva;
 import com.msduoc.reservadehoteles.models.Reserva;
+import com.msduoc.reservadehoteles.repository.ClienteRepository;
+import com.msduoc.reservadehoteles.repository.HabitacionRepository;
 import com.msduoc.reservadehoteles.repository.ReservaRepository;
+import com.msduoc.reservadehoteles.models.Cliente;
+import com.msduoc.reservadehoteles.models.Habitacion;
 
 @Service
 public class ReservaServiceImpl implements ReservaService{
     @Autowired
     private ReservaRepository reservaRepository;
 
+    @Autowired
+    private ClienteRepository clienteRepository;
+
+    @Autowired
+    private HabitacionRepository habitacionRepository;
+
     @Override
     public List<Reserva> getAllReservas() {
+        if (reservaRepository.findAll().isEmpty()) {
+            throw new RuntimeException("/GET No hay ninguna reserva.");
+        }
         return reservaRepository.findAll();
     }
 
     @Override 
     public Optional<Reserva> getReservaById(Long id) {
-        return reservaRepository.findById(id);
+        if (reservaRepository.findById(id) != null) {
+            return reservaRepository.findById(id);
+        } else {
+            throw new RuntimeException("/GET id de reserva no encontrado para mostrar.");
+        }
     }
 
     @Override
     public Reserva createReserva(Reserva reserva) {
+        // El request me está entregando la habitación?
+        if (reserva.getHabitacion() == null || reserva.getHabitacion().getId() == null) {
+            throw new RuntimeException("/POST Debe especificar una habitación válida.");
+        }
+
+        // El request me está entregando el cliente?
+        if (reserva.getCliente() == null || reserva.getCliente().getId() == null) {
+            throw new RuntimeException("/POST Debe especificar un cliente válido.");
+        }
+
+        Long clienteId = reserva.getCliente().getId();
+        Long habitacionId = reserva.getHabitacion().getId();
+
+        Cliente cliExistente = clienteRepository.findById(clienteId).get();
+        Habitacion habExistente = habitacionRepository.findById(habitacionId).get();
+
+        // Mostrar datos de otras entidades
+        reserva.setCliente(cliExistente);
+        reserva.setHabitacion(habExistente);
+
+        // Automatizar fecha y estado de reserva
+        reserva.setFechaReserva(LocalDate.now());
+        reserva.setEstado(EstadoReserva.PENDIENTE);
+
         return reservaRepository.save(reserva);
     }
 
@@ -59,12 +103,41 @@ public class ReservaServiceImpl implements ReservaService{
 
             return reservaRepository.save(reserva);
         } else {
-            return null;
+            throw new RuntimeException("/PUT id de reserva no encontrado para actualizar.");
         }
     }
 
     @Override
     public void deleteReserva(Long id) {
-        reservaRepository.deleteById(id);
+        if (reservaRepository.findById(id) != null) {
+            reservaRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("/DELETE id de reserva no encontrado para eliminar.");
+        }
+    }
+
+    @Override
+    public Reserva cancelarReserva(Long id) {
+        if (reservaRepository.findById(id) != null) {
+            Reserva res = reservaRepository.findById(id).get();
+            res.setEstado(EstadoReserva.CANCELADA);
+            return res;
+        } else {
+            throw new RuntimeException("/PUT id de reserva no encontrado para cancelar.");
+        }
+    }
+
+    @Override
+    public List<Reserva> findByEstado(EstadoReserva estado) {
+        List<Reserva> todas = reservaRepository.findAll();
+        List<Reserva> filtradas = new ArrayList<>();
+
+        for (Reserva res : todas) {
+            if (res.getEstado().equals(estado)) {
+                filtradas.add(res);
+            }
+        }
+
+        return filtradas;
     }
 }
